Update patch set 3
Patch Set 3:

The reason this is failing is because you are ifdefing out the src_java and gen_src_java definitions, and then building the java library, which calls find_all_java files or whatever.
On Mac, the path argument is required. On Linux, it is not:

What is being called:
find $(src_java) -name '*.java'

What that is being translated to:

find -name '*.java'

There is no path argument because it is ifdef'd out. Your change ifdefs out entirely more than is necessary. Revert all your commits, and ONLY ifdef the include $(BUILD*) directives.

Patch-set: 3