Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The idea sounds good. But what if instead of hard-coding the libril/ path into the makefile, you use TARGET_LIBRIL_PATH=$(LOCAL_PATH)/libril and then just check for the existence of the variable and build it instead of the default one?

Patch-set: 1