Update patch set 1
Patch Set 1:

Because the recursive makefile inclusion is voodoo, and you risk ending with "libril is already defined" errors. We had a ton of those recently when we applied the findleaves patch, and I still haven't found a safe and documented way of making sure that if you have 2 makefiles for the same module, only the one you want is included. This is particularly true for stuff that's inside device directories.

But by all means, if you have a cleaner suggestion (test it first by building different devices ;) ), please go go ahead. This particular implementation was tested by building the ONE, the z71 (both of which have their own librils) and bravo (which uses stock), without any conflicts.

Patch-set: 1