Update patch set 1
Patch Set 1:

sorry, in the end i could test this, as its libril.so, not rild. I've patched it up and did a bit of testing with RIL (sleep/calls/network), and it seems to be okay, tho i can't really compare the sleep times. Seems to not hurt stuff on my device.

Patch-set: 1