Update patch set 1
Patch Set 1: Code-Review-2

Need a better name for the symbol.  I'm leaning toward making a global symbol such as TARGET_VENDOR_BSP_VERSION which would correspond to the Android version that blobs are from.  This could be picked up in the Android.mk files and used to set a macro in CFLAGS of the same or similar name.

Patch-set: 1
Label: Code-Review=-2