Update patch set 1
Patch Set 1: Code-Review+1

Needed for

http://review.cyanogenmod.org/#/c/121249/

which is fixing the fact that M source pretty much requires that you build libril (it exports headers that are then needed to build rild).  Dan worked around that by letting the build system both build libril and install it as a proprietary file.

This worked well for him, but failed miserably on my build system (he gets the proprietary version, I get the source built version).

This is needed to allow us to use a proprietary rild to match our proprietary libril.

Patch-set: 1
Label: Code-Review=+1