Update patch set 1
Patch Set 1: Code-Review+1

This seems right to me that since the set is command #140 and get is command #141 and set (#140) requires ril >= 14 then get (#141) should also require it.

I looked at this because on pme (HTC 10) this request generates no response from our vendor blobs and causes Settings >> Developer Options to hang.

Throwing this up here to see if you guys agree and then we can merge it.

It actually doesn't matter a whole lot to me as I'm going to need to fork a custom libril for pme to handle SignalStrength changes in our blobs, but I feel like this is probably right and probably an improvement.

Patch-set: 1
Label: Code-Review=+1