ril: Setting ril to v10 or v11 serves no purpose
For the purposes of building libril, there is no difference in setting
RIL_VERSION and LAST_IMPRECISE_RIL_VERSION to any of 10, 11, or 12, so
long as the version is greater or equal to that of the vendor provided
ril blobs. Accept the default of 12.

Revert "ril-caf: Allow using RIL version 10"

This reverts commit b0d6cd340000a751d41a9c1dd83c6a9aaed8b2d9.

Revert "ril: Add support for selecting RIL version 11"

This reverts commit f77a171b5a0042a85ff0b4830d70147751f2f852.

Change-Id: I488d3cedb245521705177cddeaf3ae58bbbc6f44