Update patch set 1
Patch Set 1: Code-Review+1

There really is a difference in setting this and this change could break devices.

For example, in responseCellInfoList it reads responseCellInfoListV12 for all (callbacks) versions >= 12 and V6 for version < 10.  If you had defined the last imprecise ril version to be 10 then it would read responseCellInfoListV12 for all (callbacks) version > 10.

That said, I would say go for it and if it breaks anything we should fix it properly.  Setting this option is pretty magical and I would have no idea why it might be needed on victara and I'd rather have it break and end up having it fixed in a way that makes sense instead of just some random magic.

Patch-set: 1
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1